requirements-dev.txt
django-extensions==2.1.6
django-querycount-0.7.0
